+++ MetroidDeOuf/Cours04/Player.hpp
@@ -72,7 +72,7 @@ public:
 	void jumpBehaviour();
 
 	void checkIfInDeathZone();
-	void takeDamages(float rawDamages, int otherXX = 0, int otherYY = 0, int _knockbackForce = 0);
+	void takeDamages(int rawDamages, int otherXX = 0, int otherYY = 0, int _knockbackForce = 0);
 	void heal(float rawHeal);
 	void kill();
